refined API
